Modify eslint rule for void Promise return type
In order to use 'await' inside a controller passed to a route, the
controller must be made an async function. The issue is that this
results in a returned 'Promise<...>' type. Calling the main function
in 'src/index.ts' at top level means that it needs to be marked with
'void' so that the 'Promise<...>' returned is discarded. Therefore, the
rule must be given an additional option that permits statements that both
return a 'Promise' and are marked 'void'.