Fix runtype mismatch
When testing or retrieving data in the frontend, the date field
is a string type but when retrieving data from the database, the
date field is a Date object. To use the same type definition
uniformly, the type definition was changed to a union of a string
type and Date type.